+++ src/postgres/teams/team.rs
@@ -18,10 +18,10 @@ impl PostgresTeamRepository {
 
 #[derive(FromRow)]
 struct TeamRecord {
-    id: i32,
+    id: i64,
     name: String,
     slug: String,
-    owner_id: i32,
+    owner_id: i64,
     created_at: DateTime<Utc>,
     updated_at: DateTime<Utc>,
 }
@@ -29,10 +29,10 @@ struct TeamRecord {
 impl From<TeamRecord> for Team {
     fn from(row: TeamRecord) -> Self {
         Team {
-            id: row.id,
+            id: row.id as u64,
             name: row.name,
             slug: row.slug,
-            owner_id: row.owner_id,
+            owner_id: row.owner_id as u64,
             created_at: row.created_at,
             updated_at: row.updated_at,
         }
@@ -52,7 +52,7 @@ impl TeamRepository for PostgresTeamRepository {
         )
         .bind(&data.name)
         .bind(&data.slug)
-        .bind(data.owner_id)
+        .bind(data.owner_id as i64)
         .fetch_one(&self.pool)
         .await
         .map_err(|e| {
@@ -64,11 +64,11 @@ impl TeamRepository for PostgresTeamRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn find_by_id(&self, id: i32) -> Result<Option<Team>, AuthError> {
+    async fn find_by_id(&self, id: u64) -> Result<Option<Team>, AuthError> {
         let row: Option<TeamRecord> = sqlx::query_as(
             "SELECT id, name, slug, owner_id, created_at, updated_at FROM teams WHERE id = $1",
         )
-        .bind(id)
+        .bind(id as i64)
         .fetch_optional(&self.pool)
         .await
         .map_err(|e| {
@@ -98,7 +98,7 @@ impl TeamRepository for PostgresTeamRepository {
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
     async fn update(
         &self,
-        id: i32,
+        id: u64,
         name: Option<&str>,
         slug: Option<&str>,
     ) -> Result<Team, AuthError> {
@@ -113,7 +113,7 @@ impl TeamRepository for PostgresTeamRepository {
                 )
                 .bind(n)
                 .bind(s)
-                .bind(id)
+                .bind(id as i64)
                 .fetch_one(&self.pool)
                 .await
             }
@@ -126,7 +126,7 @@ impl TeamRepository for PostgresTeamRepository {
                     ",
                 )
                 .bind(n)
-                .bind(id)
+                .bind(id as i64)
                 .fetch_one(&self.pool)
                 .await
             }
@@ -139,7 +139,7 @@ impl TeamRepository for PostgresTeamRepository {
                     ",
                 )
                 .bind(s)
-                .bind(id)
+                .bind(id as i64)
                 .fetch_one(&self.pool)
                 .await
             }
@@ -147,7 +147,7 @@ impl TeamRepository for PostgresTeamRepository {
                 sqlx::query_as(
                     "SELECT id, name, slug, owner_id, created_at, updated_at FROM teams WHERE id = $1",
                 )
-                .bind(id)
+                .bind(id as i64)
                 .fetch_one(&self.pool)
                 .await
             }
@@ -164,9 +164,9 @@ impl TeamRepository for PostgresTeamRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn delete(&self, id: i32) -> Result<(), AuthError> {
+    async fn delete(&self, id: u64) -> Result<(), AuthError> {
         sqlx::query("DELETE FROM teams WHERE id = $1")
-            .bind(id)
+            .bind(id as i64)
             .execute(&self.pool)
             .await
             .map_err(|e| {
@@ -178,11 +178,11 @@ impl TeamRepository for PostgresTeamRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn find_by_owner(&self, owner_id: i32) -> Result<Vec<Team>, AuthError> {
+    async fn find_by_owner(&self, owner_id: u64) -> Result<Vec<Team>, AuthError> {
         let rows: Vec<TeamRecord> = sqlx::query_as(
             "SELECT id, name, slug, owner_id, created_at, updated_at FROM teams WHERE owner_id = $1 ORDER BY created_at DESC",
         )
-        .bind(owner_id)
+        .bind(owner_id as i64)
         .fetch_all(&self.pool)
         .await
         .map_err(|e| {
@@ -194,7 +194,7 @@ impl TeamRepository for PostgresTeamRepository {
     }
 
     #[cfg_attr(feature = "tracing", tracing::instrument(skip(self), err))]
-    async fn transfer_ownership(&self, team_id: i32, new_owner_id: i32) -> Result<Team, AuthError> {
+    async fn transfer_ownership(&self, team_id: u64, new_owner_id: u64) -> Result<Team, AuthError> {
         let row: TeamRecord = sqlx::query_as(
             r"
             UPDATE teams SET owner_id = $1, updated_at = NOW()
@@ -202,8 +202,8 @@ impl TeamRepository for PostgresTeamRepository {
             RETURNING id, name, slug, owner_id, created_at, updated_at
             ",
         )
-        .bind(new_owner_id)
-        .bind(team_id)
+        .bind(new_owner_id as i64)
+        .bind(team_id as i64)
         .fetch_one(&self.pool)
         .await
         .map_err(|e| match e {
